Restructure clipper
